requirements.txt
fastapi
uvicorn
tensorflow==2.12.0
h5py==3.8.0
sqlalchemy
psycopg2-binary
pydantic
numpy
watchdog
scikit-learn
pandas
